Always allow accounts with always_dispatch to connect
The always_dispatch account setting historically would allow an account
to connect without regard to connectivity, which is needed by the ring
manager. Kludge-transport did not obey that setting, and would interfere
with connections on those accounts.

This allows the ring manager to coexist with a connectivity backend, and
maintain its connection even when not connected to the internet.